Fix performance problem in regression test #14
After tracing through bsdiff to debug the performance issue seen in
regression test #14, I noticed that several conditions are present that,
combined, result in the performance hit:

  - exact matches are found at the same position in the old file
    multiple times in a row
  - the length of the exact match is one less each time
  - the matches are very similar to the bytes in the old file at the
    last offset, and the match length is very large

Short circuit the detection of a mismatched-by-8-bytes section when this
situation is encountered.

FIXME: match length of 500 is arbitrary... testing should reveal a
better threshold.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>